修改主键及配置文件主键类型

+++ service-projects/promotion-service/src/main/java/com/winhxd/b2c/promotion/dao/CouponGradeMapper.java
@@ -3,13 +3,13 @@ package com.winhxd.b2c.promotion.dao;
 import com.winhxd.b2c.common.domain.promotion.model.CouponGrade;
 
 public interface CouponGradeMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(CouponGrade record);
 
     int insertSelective(CouponGrade record);
 
-    CouponGrade selectByPrimaryKey(Integer id);
+    CouponGrade selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(CouponGrade record);
