+++ service-projects/promotion-service/src/main/java/com/winhxd/b2c/promotion/dao/CouponInvestorDetailMapper.java
@@ -3,13 +3,13 @@ package com.winhxd.b2c.promotion.dao;
 import com.winhxd.b2c.common.domain.promotion.model.CouponInvestorDetail;
 
 public interface CouponInvestorDetailMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(CouponInvestorDetail record);
 
     int insertSelective(CouponInvestorDetail record);
 
-    CouponInvestorDetail selectByPrimaryKey(Integer id);
+    CouponInvestorDetail selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(CouponInvestorDetail record);
